Create ConcreteTorchvisionBackbone class for reuse in mobilenet and efficientnet backbones
